Properly handle nodata value
Specifying 'None' as a nodata value works for native float32 format datasets such as glo_30. But for int16 format datasets such as SRTM, gdal does not handle 'None' as a nodata value, such that the native integer nodata values (-32768) in the array must be masked to avoid interpolation and geoid removal artifacts.